requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
requests 